Switch to structopt for CLI parsing, and change path flag to positional argument
Switch from clap to structopt for the CLI library of choice. This is
because we were using unversioned clap at its latest main branch
release. Structopt will provide us with concrete version support until
clap v3 releases, or we stay with structopt. We also change the
"-p/--path" flag to an optional, positional argument with this commit.