Switch back to xtask and add mangen
This change switches back to xtask and adds the ability to generate
manpages. As a result of these changes, there are now four total crates
instead of one (once again, this repository is using cargo workspaces).
Only "gfold" is able to be published and "xtask" replaces "just", but
the "cli" and "config" crates were needed to for the new "mangen" task.
The "gfold" crate is the default member, which helps package maintainers
since all automation should remain the same.

All commands made it over except for the following...

- "build" (debug) --> wasn't used
- "udeps" --> required nightly and wasn't often used
- "install-deps" --> the script is now gone

The "size" target uses native Rust logic instead of binaries on the
host. All other targets are identically executed.

Additional changes:

- macOS 13 x86_64 runner images have been replaced with "macos-15-intel"
  images, which will be supported until Fall 2027
  (https://github.blog/changelog/2025-09-19-github-actions-macos-13-runner-image-is-closing-down/)
- The "install-deps" script has been removed
- References to "just" commands and usage in CI and docs have been
  removed or replaced
- Fail docs check on warnings

Signed-off-by: Nick Gerace <nickagerace@gmail.com>