Add flag to generate a man page
This change adds a flag to generate a man page. This was done by an
xtask, but now is done by the CLI itself. This is how ripgrep works, for
what it's worth. The xtask still generates the man page, but now it
does it via "cargo run" and pulls it up using the "man" command.

In addition, all supporting crates for gfold have been collapsed back
into one. When publishing gfold, even with "publish = false", supporting
crates need to be versioned. Since the generation of man pages is now
supported via a flag, xtask no longer needs to depend on a gfold
library anyway.

Signed-off-by: Nick Gerace <nickagerace@gmail.com>